Move internal data wrappers to model/internal
